# backend/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: Directory Wizard API
  description: |
    Multi-tenant Directory Creation Wizard API for creating and managing directory websites
    through a guided interface with integrated external service provisioning.
  version: 1.0.0
  contact:
    name: Directory Wizard Support
    email: support@directorywizard.com
    url: https://docs.directorywizard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.directorywizard.com/api
    description: Production server
  - url: https://staging-api.directorywizard.com/api
    description: Staging server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /tenants:
    post:
      summary: Create tenant
      description: Create a new tenant (directory site)
      tags:
        - Tenants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantCreateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /tenants/{id}:
    get:
      summary: Get tenant
      description: Retrieve tenant details by ID
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update tenant
      description: Update tenant basic information
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantUpdateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tenants/{id}/branding:
    put:
      summary: Update tenant branding
      description: Update tenant branding and theme with file uploads
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrandingRequest'
      responses:
        '200':
          description: Branding updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '413':
          $ref: '#/components/responses/FileTooLargeError'
  /tenants/{id}/upload:
    post:
      summary: Upload data file
      description: Upload categories or listings data file
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [categories, listings]
          description: Type of data file to upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          description: File uploaded and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '413':
          $ref: '#/components/responses/FileTooLargeError'
        '422':
          $ref: '#/components/responses/FileValidationError'
  /tenants/{id}/preview:
    get:
      summary: Get tenant preview
      description: Generate tenant preview with statistics and structure
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: format
          in: query
          schema:
            type: string
            enum: [json, html, template, metadata]
            default: json
          description: Response format
        - name: include_content
          in: query
          schema:
            type: boolean
            default: false
          description: Include full content in response
        - name: include_drafts
          in: query
          schema:
            type: boolean
            default: false
          description: Include draft content
        - name: max_listings
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum number of listings to include
      responses:
        '200':
          description: Preview generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewResponse'
            text/html:
              schema:
                type: string
                description: HTML preview (when format=html)
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tenants/{id}/publish:
    post:
      summary: Publish tenant
      description: Start tenant publishing process
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '202':
          description: Publishing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /tenants/{id}/jobs/{jobId}:
    get:
      summary: Get job status
      description: Get publishing job status and progress
      tags:
        - Tenants
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^tnt_[a-zA-Z0-9]+$'
      description: Tenant ID
      example: tnt_123456789
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        pattern: '^job_[a-zA-Z0-9]+$'
      description: Job ID
      example: job_123456789
...
  schemas:
...
    CreateTenantRequest:
      type: object
      required:
        - name
        - domain
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: San Francisco Business Directory
        domain:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
          example: sf-business-directory
    TenantCreateResponse:
      type: object
      properties:
        id:
          type: string
          example: tnt_123456789
        name:
          type: string
          example: San Francisco Business Directory
        domain:
          type: string
          example: sf-business-directory
        status:
          type: string
          enum: [DRAFT, UPDATING, PUBLISHED, FAILED]
          example: DRAFT
        session_id:
          type: string
          example: ses_987654321
        next_step:
          type: string
          example: BRANDING
        created_at:
          type: string
          format: date-time
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Updated Directory Name
        domain:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
          example: updated-domain
    TenantUpdateResponse:
      type: object
      properties:
        id:
          type: string
          example: tnt_123456789
        name:
          type: string
          example: Updated Directory Name
        domain:
          type: string
          example: updated-domain
        status:
          type: string
          enum: [DRAFT, UPDATING, PUBLISHED, FAILED]
          example: DRAFT
        updated_at:
          type: string
          format: date-time
    TenantResponse:
      type: object
      properties:
        id:
          type: string
          example: tnt_123456789
        name:
          type: string
          example: San Francisco Business Directory
        domain:
          type: string
          example: sf-business-directory
        status:
          type: string
          enum: [DRAFT, UPDATING, PUBLISHED, FAILED]
          example: PUBLISHED
        branding:
          $ref: '#/components/schemas/Branding'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
        listings:
          type: array
          items:
            $ref: '#/components/schemas/ListingSummary'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
    Branding:
      type: object
      properties:
        logo_url:
          type: string
          format: uri
          example: https://storage.example.com/logos/logo.png
        primary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#3B82F6'
        secondary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#1F2937'
        accent_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#F59E0B'
        font_family:
          type: string
          example: Inter
        font_url:
          type: string
          format: uri
          example: https://fonts.google.com/inter
    CategorySummary:
      type: object
      properties:
        id:
          type: string
          example: cat_123
        name:
          type: string
          example: Restaurants
        slug:
          type: string
          example: restaurants
        listings_count:
          type: integer
          example: 25
...
    BrandingRequest:
      type: object
      required:
        - primary_color
        - secondary_color
        - accent_color
        - font_family
      properties:
        primary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#3B82F6'
        secondary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#1F2937'
        accent_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#F59E0B'
        font_family:
          type: string
          example: Inter
        logo:
          type: string
          format: binary
          description: Logo image file (PNG, JPG, SVG, max 5MB)
        font_file:
          type: string
          format: binary
          description: Font file (TTF, OTF, WOFF, WOFF2, max 2MB)
        session_id:
          type: string
          example: ses_987654321
    BrandingResponse:
      type: object
      properties:
        tenant_id:
          type: string
          example: tnt_123456789
        branding:
          $ref: '#/components/schemas/Branding'
        next_step:
          type: string
          example: CATEGORIES
    FileUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: Data file (CSV, JSON, max 10MB)
    FileUploadResponse:
      type: object
      properties:
        file_id:
          type: string
          example: fil_123456789
        type:
          type: string
          enum: [categories, listings]
          example: categories
        filename:
          type: string
          example: categories.json
        records_count:
          type: integer
          example: 5
        validation_status:
          type: string
          enum: [VALID, INVALID, WARNING]
          example: VALID
        next_step:
          type: string
          example: LISTINGS
    PreviewResponse:
      type: object
      properties:
        tenant_id:
          type: string
          example: tnt_123456789
        name:
          type: string
          example: San Francisco Business Directory
        domain:
          type: string
          example: sf-business-directory
        status:
          type: string
          example: DRAFT
        preview_url:
          type: string
          format: uri
          example: https://sf-business-directory.example.com/preview
        admin_url:
          type: string
          format: uri
          example: https://admin.example.com/tenant/tnt_123456789
        statistics:
          $ref: '#/components/schemas/Statistics'
        branding:
          $ref: '#/components/schemas/Branding'
        site_structure:
          $ref: '#/components/schemas/SiteStructure'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryPreview'
        listings:
          type: array
          items:
            $ref: '#/components/schemas/ListingPreview'
        readiness:
          $ref: '#/components/schemas/ReadinessCheck'
...
    PublishResponse:
      type: object
      properties:
        message:
          type: string
          example: Publishing started
        job_id:
          type: string
          example: job_123456789
        tenant_id:
          type: string
          example: tnt_123456789
        status:
          type: string
          enum: [QUEUED, IN_PROGRESS, COMPLETED, FAILED, CANCELLED]
          example: QUEUED
        estimated_duration:
          type: string
          example: 2-5 minutes
    JobStatusResponse:
      type: object
      properties:
        job_id:
          type: string
          example: job_123456789
        tenant_id:
          type: string
          example: tnt_123456789
        status:
          type: string
          enum: [QUEUED, IN_PROGRESS, COMPLETED, FAILED, CANCELLED]
          example: IN_PROGRESS
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 65
        steps:
          type: array
          items:
            $ref: '#/components/schemas/JobStep'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Detailed error description
        field:
          type: string
          example: email
        code:
          type: string
          example: VALIDATION_ERROR
        timestamp:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    FileTooLargeError:
      description: File size exceeds limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    FileValidationError:
      description: File content validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrors'
...